Mine Text Components in the INLPO Publications

Performs a word frequency text analysis of Idaho National Laboratory Project Office (INLPO) publications.

MineText(
  pubs,
  components = c("title", "abstract"),
  ngmin = 1L,
  ngmax = ngmin,
  lowfreq = 1L
)

Arguments

pubs

pubs_data class. Bibliographic information, see pubs dataset for details.

components

character vector. One or more text components to analyze. Choices include the "title", "abstract", "annotation", and "citation" of the document.

ngmin, ngmax

integer number. Splits strings into n-grams with given minimal and maximal numbers of grams. An n-gram is an ordered sequence of n words taken from the body of a text. Requires that the RWeka package is available and that the environment variable JAVA_HOME points to where the Java software is located. Recommended for single text compoents only.

lowfreq

integer number. Lower frequency bound. Words that occur less than this bound are excluded from the returned frequency table.

Value

A word frequency table giving the number of times each word occurs in a publication's text component(s). A table column represents a single publication that is identified using its citation-key. And each row provides frequency counts for a particular word (also known as a ‘term’).

Details

HTML entities are decoded when the textutils package is available.

See also

Author

J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center

Examples

m <- MineText(head(pubs, 3))
head(m)
#>                Docs
#> Terms           Fisher2022 BartholomayOthers2021 Fisher2021
#>   idaho                  6                     4          1
#>   quality                0                     7          1
#>   aquifer                2                     0          5
#>   national               3                     3          0
#>   water-quality          0                     3          3
#>   laboratory             3                     2          0
if (FALSE) {
d <- data.frame(word = rownames(m), freq = rowSums(m))
inlmisc::MakeWordCloud(d, display = TRUE)
}