Bibliographic Information of the INLPO

Bibliographic information for reports, articles, maps, and theses related to scientific monitoring and research conducted by the U.S. Geological Survey (USGS), Idaho Water Science Center, Idaho National Laboratory Project Office (INLPO).

pubs

Format

An object of class ‘pubs_data’ that inherits behavior from the data frame class. Each record corresponds to a bibliographical item and contains the following variables:

key

BibTeX key for the citation entry;

year

year of publication;

citation

bibliographic entry of class bibentry;

abstract

abstract text string;

annotation

Knobel and others (2005) annotation text string.

Row names are the BibTeX key for the citation entry.

Source

Many of these publications are available through the USGS Publications Warehouse.

References

Knobel, L.L., Bartholomay, R.C., and Rousseau, J.P., 2005, Historical development of the U.S. Geological Survey hydrologic monitoring and investigative programs at the Idaho National Engineering and Environmental Laboratory, Idaho, 1949 to 2001: U.S. Geological Survey Open-File Report 2005--1223 (DOE/ID--22195), 93 p., doi: 10.3133/ofr20051223 .

Examples

## Display table structure
str(pubs, max.level = 1, nchar.max = 50)
#> Classes ‘pubs_data’ and 'data.frame':	366 obs. of  5 variables:
#>  $ key       : chr  "Fisher2022" "BartholomayOthers2021" "Fisher2021" "FisherOthers2021" ...
#>  $ year      : int  2022 2021 2021 2021 2021 2021 2020 2020 2020 2020 ...
#>  $ citation  :List of 366
#>   ..- attr(*, "class")= chr [1:2] "AsIs" "bibentry"
#>  $ abstract  : chr  "The R package inlpubs may be used"| __truncated__ "Water-quality activities and wate"| __truncated__ "The establishment of an efficient"| __truncated__ "Long-term monitoring of water-qua"| __truncated__ ...
#>  $ annotation: chr  NA NA NA NA ...

## Print the citation key for each entry in the bibliography:
rownames(pubs[1:10, ])
#>  [1] "Fisher2022"            "BartholomayOthers2021" "Fisher2021"           
#>  [4] "FisherOthers2021"      "TwiningOthers2021a"    "TwiningOthers2021b"   
#>  [7] "BartholomayOthers2020" "Fisher2020"            "HelmuthOthers2020"    
#> [10] "SchuslerOthers2020"   

## Print citation, authors, and abstract for Fisher and others (2012):
key <- "FisherOthers2012"
ref <- pubs[key, "citation"]
print(ref, style = "citation", bibtex = TRUE)
#> 
#>   Fisher, J.C., Rousseau, J.P., Bartholomay, R.C, and Rattray, G.W.,
#>   2012, A comparison of U.S. Geological Survey three-dimensional model
#>   estimates of groundwater source areas and velocities to independently
#>   derived estimates, Idaho National Laboratory and vicinity, Idaho:
#>   U.S. Geological Survey Scientific Investigations Report 2012--5152
#>   (DOE/ID--22218), 130 p., https://doi.org/10.3133/sir20122152.
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @TechReport{FisherOthers2012,
#>     title = {A comparison of U.S. Geological Survey three-dimensional model estimates of groundwater source areas and velocities to independently derived estimates, Idaho National Laboratory and vicinity, Idaho},
#>     author = {Jason C. Fisher and Joseph P. Rousseau and Roy C. Bartholomay and Gordon W. Rattray},
#>     institution = {U.S. Geological Survey},
#>     type = {Scientific Investigations Report},
#>     year = {2012},
#>     number = {2012--5152 (DOE/ID--22218)},
#>     pages = {130},
#>     doi = {10.3133/sir20122152},
#>   }
#> 
format(ref$author, c("given", "family"))
#>              jfisher            jrousseau         rbartholomay 
#>    "Jason C. Fisher" "Joseph P. Rousseau" "Roy C. Bartholomay" 
#>             grattray 
#>  "Gordon W. Rattray" 
txt <- pubs[key, "abstract"]
cat(strwrap(txt), sep = "\n")
#> The U.S. Geological Survey (USGS), in cooperation with the U.S.
#> Department of Energy, evaluated a three-dimensional model of
#> groundwater flow in the fractured basalts and interbedded sediments of
#> the eastern Snake River Plain aquifer at and near the Idaho National
#> Laboratory to determine if model-derived estimates of groundwater
#> movement are consistent with (1) results from previous studies on water
#> chemistry type, (2) the geochemical mixing at an example well, and (3)
#> independently derived estimates of the average linear groundwater
#> velocity. Simulated steady-state flow fields were analyzed using
#> backward particle-tracking simulations that were based on a modified
#> version of the particle tracking program MODPATH. Model results were
#> compared to the 5-microgram-per-liter lithium contour interpreted to
#> represent the transition from a water type that is primarily composed
#> of tributary valley underflow and streamflow-infiltration recharge to a
#> water type primarily composed of regional aquifer water. This
#> comparison indicates several shortcomings in the way the model
#> represents flow in the aquifer. The eastward movement of tributary
#> valley underflow and streamflow-infiltration recharge is overestimated
#> in the north-central part of the model area and underestimated in the
#> central part of the model area. Model inconsistencies can be attributed
#> to large contrasts in hydraulic conductivity between hydrogeologic
#> zones.\n\nSources of water at well NPR-W01 were identified using
#> backward particle tracking, and they were compared to the relative
#> percentages of source water chemistry determined using geochemical mass
#> balance and mixing models. The particle tracking results compare
#> reasonably well with the chemistry results for groundwater derived from
#> surface-water sources (\u201328 percent error), but overpredict the
#> proportion of groundwater derived from regional aquifer water (108
#> percent error) and underpredict the proportion of groundwater derived
#> from tributary valley underflow from the Little Lost River valley
#> (\u201374 percent error). These large discrepancies may be attributed
#> to large contrasts in hydraulic conductivity between hydrogeologic
#> zones and (or) a short-circuiting of underflow from the Little Lost
#> River valley to an area of high hydraulic
#> conductivity.\n\nIndependently derived estimates of the average
#> groundwater velocity at 12 well locations within the upper 100 feet of
#> the aquifer were compared to model-derived estimates. Agreement between
#> velocity estimates was good at wells with travel paths located in areas
#> of sediment-rich rock (root-mean-square error [RMSE] = 5.2 feet per day
#> [ft/d]) and poor in areas of sediment-poor rock (RMSE = 26.2 ft/d);
#> simulated velocities in sediment-poor rock were 2.5 to 4.5 times larger
#> than independently derived estimates at wells USGS 1 (less than 14
#> ft/d) and USGS 100 (less than 21 ft/d). The models overprediction of
#> groundwater velocities in sediment-poor rock may be attributed to large
#> contrasts in hydraulic conductivity and a very large, model-wide
#> estimate of vertical anisotropy (14,800).

## Print list of authors:
authors <- do.call("c", pubs$citation$author)
authors <- authors[!duplicated(authors)]
cat(format(authors[1:10]), sep = "\n")
#> Jason C. Fisher <jfisher@usgs.gov> (<https://orcid.org/0000-0001-9032-8912>)
#> Roy C. Bartholomay <rcbarth@usgs.gov> (<https://orcid.org/0000-0002-4809-9287>)
#> Neil V. Maimer <nmaimer@usgs.gov> (<https://orcid.org/0000-0003-3047-3282>)
#> Amy J. Wehnke <ajwehnke@usgs.gov>
#> Samuel L. Helmuth <shelmuth@usgs.gov> (<https://orcid.org/0000-0002-2880-7290>)
#> Gordon W. Rattray <grattray00@gmail.com> (<https://orcid.org/0000-0002-1690-3218>)
#> Brian V. Twining <btwining@usgs.gov> (<https://orcid.org/0000-0003-1321-4721>)
#> Calvin Anderson
#> Blair W. Packer (<https://orcid.org/0000-0002-6234-9115>)
#> Evan J. Martin

if (FALSE) {
## Export suggested citations from the bibliography:
txt <- sort(vapply(pubs$citation, function(x) {
  attr(unclass(x)[[1]], "textVersion")
}, character(1)))
txt <- head(c(rbind(txt, character(length(txt)))), -1)
txt <- strwrap(txt, width = 80, exdent = 2)
file <- tempfile(fileext = ".txt")
writeLines(txt, file)
file.show(file, title = "Suggested citations")
}